Check if one delimited string is a subset of another (case-insensitive and whitespace-trimmed)

For each pair of corresponding strings in a and b, applies OR.delim.split() and checks whether all elements in b are contained in a.

Usage

OR.delim.subset(a, b, delimiter = ",")

Arguments

a

String vector, matrix, or data frame.

b

String vector, matrix, or data frame of the same shape as a.

delimiter

String delimiter. Default = ",".

Value

A logical vector or matrix indicating whether b is a subset of a.

See also