Cox regression with time-dependent covariates

Fits a Cox proportional-hazards model with one or more binary time-dependent covariates, each taking the value 0 before its transition time and 1 thereafter. Follow-up is split at each transition time, so that time at risk preceding the transition is correctly assigned to the unexposed state, avoiding immortal time bias.

Usage

OR.tdCox(id, time, status, X.tdc, X = NULL)

Arguments

id

Vector of subject identifiers.

time

Numeric vector of follow-up times.

status

Numeric vector of event indicators, where 1 denotes the event and 0 denotes censoring.

X.tdc

Data frame of transition times for each time-dependent covariate, with NA where the transition does not occur. Column names are used as covariate names.

X

Optional data frame of time-independent covariates. Default = NULL.

Value

An object of class coxph.

See also

Examples

set.seed(24601)
n <- 500
data <- data.frame("id" = 1:n)
data$date.event <- rexp(n, rate = log(2)/6)
data$date.censor <- rexp(n, rate = log(2)/12)
data$time <- pmin(data$date.event, data$date.censor)
data$status <- as.numeric(data$date.event <= data$date.censor)
data$date.cycle1 <- 1
data$date.cycle1[runif(n) < 0.05] <- NA
data$date.cycle1[data$date.cycle1 > data$time] <- NA
OR.tdCox(data$id, data$time, data$status, X.tdc = data.frame("cycle1" = data$date.cycle1))
#> Call:
#> survival::coxph(formula = stats::as.formula(formula), data = data, 
#>     model = TRUE)
#> 
#>          coef exp(coef) se(coef)     z     p
#> cycle1 0.1164    1.1234   0.2446 0.476 0.634
#> 
#> Likelihood ratio test=0.23  on 1 df, p=0.6286
#> n= 897, number of events= 331